+++ tests/test_core/test_operations/test_rasterize.py
