+++ tests/test_core/test_spatialdata_operations.py
