+++ tests/test_transformations/test_ngff/test_ngff_coordinate_system.py
